18. [수정] Member -> MemberAuth 로 네이밍 변경
