+++ src/main/java/platform/ads/api/v1/member/controller/MemberController.java
@@ -1,37 +1,12 @@
 package platform.ads.api.v1.member.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import platform.ads.api.v1.member.dto.JoinDto;
-import platform.ads.api.v1.member.service.MemberService;
-import platform.ads.api.v1.member.vo.JoinVo;
-import platform.ads.config.base.Api;
-import platform.ads.config.base.ResultCode;
 
 @RestController
 @RequestMapping("${api-prefix-lv1}/member")
 @RequiredArgsConstructor
 public class MemberController {
 
-    private final MemberService memberService;
-
-    @PostMapping("/join")
-    public ResponseEntity<Api<JoinVo>> join(
-        @RequestBody @Validated JoinDto joinDto,
-        BindingResult bindingResult
-    ) {
-        return ResponseEntity.status(HttpStatus.CREATED)
-                             .body(Api.<JoinVo>builder()
-                                      .code(ResultCode.POST.code)
-                                      .message(ResultCode.POST.message)
-                                      .data(memberService.join(joinDto))
-                                      .build());
-    }
 }
